requirements.txt
torch
pandas
numpy
scikit-learn
matplotlib
seaborn
pyarrow
pyyaml
flake8
black